Feature #3422 - improve file conversion code
